+++ CHANGELOG.md
@@ -1,5 +1,7 @@
 #### CURRENT HEAD:
 * Fix slow convergence linear Davidson algorithm
+* Binary options to dump out CASPT2 intermediates (2-RDM,
+  3-RDM, and 4-RDM contracted with Fock) for MOLCAS
 
 #### Version 1.7.3 (2016-07-14):
 * DMRG workshop in sphinx documentation